QDjob_editor/requirements.txt
pycryptodome==3.20.0
pywebview==6.1
requests==2.31.0
pyinstaller==6.0.0
